Add all_workloads + ring AllReduce sweeps and watchdog-based clean shutdown
Two new sweep campaigns beyond the a3_200g baseline, with per-tier link
utilization capture. Coverage:

- all_workloads_128_1024/ -- 90 runs, 5 patterns x 3 sizes x {128,1024} x
  3 fabrics (87 ok + 3 incompletes at 8:1 incast_random_1024n, real
  transport observation).
- allreduce_ring_128_1024/ -- 9 clean 128-node ring AR runs at 3 sizes x
  3 fabrics. Ring at 1024n was attempted but OOM-killed (40-50 GB RSS per
  run -- 2M+ pre-allocated UecSrc/UecSink instances on the 213 MB matrix);
  documented in the sweep's NOTES.md, rows removed from final CSVs.

Two minimal source patches, orthogonal to transport / NSCC / CC:

- main_uec.cpp: 10-line SIGTERM handler. Sets a volatile flag; the sim
  loop breaks between events; main() epilogue runs normally
  (~Logfile transposeLog + summary print). This lets an external
  watchdog end htsim cleanly the moment matrix Connections-M flows
  finish, so we no longer have to per-workload-tune -end. Log volume
  becomes proportional to actual workload duration, not the -end cap.
- loggers.cpp: 2-line fix to QueueLoggerSampling::event_to_str for
  CUM_TRAFFIC records -- the ASCII printer cast _cumarr (seconds,
  typical 1e-6..1e-2) to int, truncating to 0 for every utilization
  sample. Binary log was fine; only the ASCII conversion needed
  the double-precision output. Enables per-tier link utilization
  extraction from parse_output.

Runner details in each sweep dir's run_one.sh + NOTES.md, and the
overview / entry point is newRuns.md at the repo root.

Excluded from these sweeps: AllReduce butterfly (reproducible bug in the
stock recv_done_trigger cascade -- only round 0 fires, deterministic
even without our logging or watchdog); All-to-All (deferred by design).
Both flagged in newRuns.md for follow-up.

+++ htsim/sim/loggers.cpp
@@ -303,8 +303,10 @@ string QueueLoggerSampling::event_to_str(RawLogEvent& event) {
         ss << " Type QUEUE_APPROX";
         ss << " ID " << event._id;
         assert(event._ev == QueueLogger::CUM_TRAFFIC);
-        ss << " Ev CUM_TRAFFIC CumArr " << (int)event._val1
-           << " CumIdle " << (int)event._val2 << " CumDrop " << (int)event._val3;
+        // CumArr/CumIdle/CumDrop are seconds-of-link-time (double, typically 1e-6..1);
+        // integer cast truncates them all to 0. Keep full precision.
+        ss << " Ev CUM_TRAFFIC CumArr " << event._val1
+           << " CumIdle " << event._val2 << " CumDrop " << event._val3;
         if (event._name!="") ss << " Name " << event._name;
         break;
     default:
